Makes builds 2x faster thanks to parallelization (except for the original single core Zero W)

- [x] Parallelize build on device. This takes a full build on a Pi5 from 90sec to 50sec. At 3.5min for Zero W2, down from 8min.
- [x] Cache parallelized build. This takes it down to 15sec on a Pi5. Under 50sec for Zero W2, down from 90 sec

We could probably save even more with cross compilation, but 1) it's cumbersome to add inside docker for all platforms, with adequate caching, etc, 2) with under minute compilation and deployment times, it doesn't really matter anymore, at least at a small scale.